flask-auth/auth/config.py
import os
from datetime import timedelta
from decouple import config

BASE_DIR = os.path.dirname(os.path.realpath(__file__))


class BaseConfig:

    SQLALCHEMY_DATABASE_URI = config("DATABASE_URL")
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SECRET_KEY = config("SECRET_KEY")
    JWT_SECRET_KEY = config("JWT_SECRET_KEY")
    JWT_ACCESS_TOKEN_EXPIRES = timedelta(hours=1)
