flask-auth/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
black==21.12b0
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
Werkzeug==2.0.2
